+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/reverse_proxy_functions/destroy_reverse_proxy.go
@@ -16,7 +16,7 @@ const (
 func DestroyReverseProxy(ctx context.Context, dockerManager *docker_manager.DockerManager) error {
 	_, maybeReverseProxyContainerId, err := getReverseProxyObjectAndContainerId(ctx, dockerManager)
 	if err != nil {
-		logrus.Warnf("Attempted to destroy reverse proxy but no reverse proxy container was found.")
+		logrus.Warnf("Attempted to destroy reverse proxy but no reverse proxy container was found. Error was:\n%s", err.Error())
 		return nil
 	}
 
@@ -24,13 +24,20 @@ func DestroyReverseProxy(ctx context.Context, dockerManager *docker_manager.Dock
 		return nil
 	}
 
-	if err := dockerManager.StopContainer(ctx, maybeReverseProxyContainerId, stopReverseProxyContainerTimeout); err != nil {
-		return stacktrace.Propagate(err, "An error occurred stopping the reverse proxy container with ID '%v'", maybeReverseProxyContainerId)
+	if err := destroyReverseProxyWithContainerId(ctx, dockerManager, maybeReverseProxyContainerId); err != nil {
+		return stacktrace.Propagate(err, "An error occurred destroying the reverse proxy container with ID '%v'", maybeReverseProxyContainerId)
 	}
 
-	if err := dockerManager.RemoveContainer(ctx, maybeReverseProxyContainerId); err != nil {
-		return stacktrace.Propagate(err, "An error occurred removing the reverse proxy container with ID '%v'", maybeReverseProxyContainerId)
+	return nil
+}
+
+func destroyReverseProxyWithContainerId(ctx context.Context, dockerManager *docker_manager.DockerManager, reverseProxyContainerId string) error {
+	if err := dockerManager.StopContainer(ctx, reverseProxyContainerId, stopReverseProxyContainerTimeout); err != nil {
+		return stacktrace.Propagate(err, "An error occurred stopping the reverse proxy container with ID '%v'", reverseProxyContainerId)
 	}
 
+	if err := dockerManager.RemoveContainer(ctx, reverseProxyContainerId); err != nil {
+		return stacktrace.Propagate(err, "An error occurred removing the reverse proxy container with ID '%v'", reverseProxyContainerId)
+	}
 	return nil
 }