+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/reverse_proxy_functions/implementations/traefik/traefik_container_config_provider.go
@@ -51,7 +51,7 @@ func (traefik *traefikContainerConfigProvider) GetContainerArgs(
 	overrideCmd := []string{
 		shCmdFlag,
 		fmt.Sprintf(
-			"%v '%v' && %v '%v' > %v && %v",
+			"%v -p '%v' && %v '%v' > %v && %v",
 			mkdirCmdName,
 			configDirpath,
 			printfCmdName,