+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend/kubernetes_kurtosis_backend.go
@@ -2,6 +2,7 @@ package kubernetes_kurtosis_backend
 
 import (
 	"context"
+	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/image_build_spec"
 	"io"
 	apiv1 "k8s.io/api/core/v1"
 
@@ -255,7 +256,7 @@ func (backend *KubernetesKurtosisBackend) StartRegisteredUserServices(
 ) {
 	restartPolicy := apiv1.RestartPolicyNever
 	if backend.productionMode {
-		restartPolicy = apiv1.RestartPolicyOnFailure
+		restartPolicy = apiv1.RestartPolicyAlways
 	}
 
 	successfullyStartedServices, failedServices, err := user_services_functions.StartRegisteredUserServices(
@@ -474,6 +475,11 @@ func (backend *KubernetesKurtosisBackend) CreateReverseProxy(ctx context.Context
 func (backend *KubernetesKurtosisBackend) DestroyReverseProxy(ctx context.Context) error {
 	// TODO IMPLEMENT
 	return stacktrace.NewError("Destroying the reverse proxy isn't yet implemented on Kubernetes")
+}
+
+func (backend *KubernetesKurtosisBackend) BuildImage(ctx context.Context, imageName string, imageBuildSpec *image_build_spec.ImageBuildSpec) (string, error) {
+	// TODO IMPLEMENT
+	return "", stacktrace.NewError("Building images isn't yet implemented in Kubernetes.")
 }
 
 // ====================================================================================================